what-movie-server/pyproject.toml
[tool.poetry]
name = "what-movie-server"
version = "0.1.0"
description = ""
authors = ["Lloyd Wang <lloydwangxy@gmail.com>"]
readme = "README.md"
packages = [{include = "what_movie_server"}]

[tool.poetry.dependencies]
python = "^3.11"
flask-cors = "^3.0.10"
flask-sqlalchemy = "^3.0.3"
flask-bcrypt = "^1.0.1"
pyjwt = "^2.7.0"
flask-testing = "^0.8.1"
flask-migrate = "^4.0.4"
coverage = "^7.2.7"
psycopg2 = "^2.9.6"
requests = "^2.31.0"
pydantic = "^1.10.9"
flask-pydantic = "^0.11.0"
flake8 = "^6.0.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
